hw1/requirements.txt
gym==0.15.4
tensorflow-gpu==2.0.0
mujoco-py==1.50.1.68
tensorflow-probability==0.8.0
tensorboardX==1.9
matplotlib==3.1.2
ipython==7.10.2
moviepy==1.0.0